refactor(app): split API endpoints into Flask blueprints and fix user API test
